Rename query-functions to fetchers
